requirements.txt
langchain==0.3.27
langchain-core==0.3.72
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
openai==1.98.0
pydantic==2.11.7
pydantic-core==2.33.2
PyPDF2>=3.0.0
python-docx>=0.8.11
python-dotenv>=1.0.0

langchain-community>=0.2.0
pandas==2.1.4
openpyxl==3.1.2
gspread==5.12.0
oauth2client==4.1.3
requests==2.31.0
typing-extensions>=4.12.2
langchain_chroma
pypdf
Flask==3.0.3